+++ docs/changelog.md
@@ -3,6 +3,10 @@
 ## v2.5.1 - Master
 ####* This version is not yet released and is under active development.
 
+###Updated
+
+  - Client
+    - Remove old archives in background thread
 
 ## v2.5.0 - 2017/09/30